+++ CHANGELOG.md
@@ -5,22 +5,36 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
-## v1.0.0 - 2016-11-17
+## [v2.0.0] - 2016-12-20
+- Simplify code
+- Drop LTS support
+- Update Code Style
+- Drop Authy deprecated sandbox api support
+- Push dependencies forward and require php7
+
+## [v1.0.0] - 2016-11-17
 - Commit first stable release
 
-## v0.0.5 - 2016-11-17
+## [v0.0.5] - 2016-11-17
 - Alias master branch for version stability requirements
 
-## v0.0.4 - 2016-11-16
+## [v0.0.4] - 2016-11-16
 - Fix wrong installation steps
 
-## v0.0.3 - 2016-11-15
+## [v0.0.3] - 2016-11-15
 - Fix wrong sensiolabs insight ID
 - Fix wrong files permissions
 - Add Travis shield
 
-## v0.0.2 - 2016-11-15
+## [v0.0.2] - 2016-11-15
 - Fix few typos
 
 ## v0.0.1 - 2016-11-15
 - Commit first draft
+
+[v2.0.0]: https://github.com/rinvex/laravel-authy/compare/v1.0.0...v2.0.0
+[v1.0.0]: https://github.com/rinvex/laravel-authy/compare/v0.0.5...v1.0.0
+[v0.0.5]: https://github.com/rinvex/laravel-authy/compare/v0.0.4...v0.0.5
+[v0.0.4]: https://github.com/rinvex/laravel-authy/compare/v0.0.3...v0.0.4
+[v0.0.3]: https://github.com/rinvex/laravel-authy/compare/v0.0.2...v0.0.3
+[v0.0.2]: https://github.com/rinvex/laravel-authy/compare/v0.0.1...v0.0.2